hibernateDatabase

+++ src/main/java/pl/coderstrust/database/HibernateInvoiceDatabase.java
@@ -37,6 +37,9 @@ public class HibernateInvoiceDatabase implements InvoiceDatabase {
     if (id == null) {
       throw new IllegalArgumentException("Id cannot be null");
     }
+    if (id < 0) {
+      throw new IllegalArgumentException("Id cannot be less than 0");
+    }
     try {
       return hibernateInvoiceRepository.findById(id);
     } catch (NoSuchElementException e) {
@@ -48,6 +51,9 @@ public class HibernateInvoiceDatabase implements InvoiceDatabase {
   public boolean existsById(Long id) throws DatabaseOperationException {
     if (id == null) {
       throw new IllegalArgumentException("Id cannot be null");
+    }
+    if (id < 0) {
+      throw new IllegalArgumentException("Id cannot be less than 0");
     }
     try {
       return hibernateInvoiceRepository.existsById(id);
@@ -78,6 +84,9 @@ public class HibernateInvoiceDatabase implements InvoiceDatabase {
   public void deleteById(Long id) throws DatabaseOperationException {
     if (id == null) {
       throw new IllegalArgumentException("Id cannot be null");
+    }
+    if (id < 0) {
+      throw new IllegalArgumentException("Id cannot be less than 0");
     }
     try {
       hibernateInvoiceRepository.deleteById(id);
